requirements.txt
langgraph==0.2.26
typing-extensions>=4.9.0
openai==1.43.0
pydantic==2.8.2
typer==0.12.5
SQLAlchemy==2.0.31
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
ruamel.yaml==0.18.6
python-dateutil==2.9.0.post0
tabulate==0.9.0
apscheduler==3.10.4
python-dotenv==1.0.1